fix: read & write multi payment count as 2 byte (spec compliance)

+++ src/Deserializers/MultiPayment.php
@@ -34,7 +34,7 @@ class MultiPayment extends AbstractDeserializer
 
         $this->transaction->asset = ['payments' => []];
 
-        $count  = $this->buffer->readUInt32() & 0xff;
+        $count  = $this->buffer->readUInt16() & 0xff;
         $offset = $this->assetOffset / 2 + 1;
 
         for ($i = 0; $i < $count; ++$i) {
